debug

+++ my_library/models/reading_comprehension/multi_granu_hierarch_attention_elmo.py
@@ -213,8 +213,8 @@ class MultiGranuFusionElmo(Model):
 		question_lstm_mask = question_mask if self._mask_lstms else None
 		passage_lstm_mask = passage_mask if self._mask_lstms else None
 
-		encoded_question = self._dropout(self._phrase_layer(embedded_question, question_lstm_mask))
-		encoded_passage = self._dropout(self._phrase_layer(embedded_passage, passage_lstm_mask))
+		encoded_question = self._dropout(self._phrase_layer(embedded_question.int(), question_lstm_mask))
+		encoded_passage = self._dropout(self._phrase_layer(embedded_passage.int(), passage_lstm_mask))
 		encoding_dim = encoded_question.size(-1)
 
 		# Shape: (batch_size, passage_length, question_length)
